% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_cxd.R
\name{add_new_cxd}
\alias{add_new_cxd}
\title{Check if data in input folder is OK and move to raw data folder}
\usage{
add_new_cxd(input, output)
}
\arguments{
\item{input}{The folder, where a folder \code{bemovi} is located which
contains the new files.}

\item{output}{A folder, which contains a subfolder called \code{bemovi}, i.e.
the usually the raw data folder, into which the fioles will be moved to.}
}
\value{
a \code{list} which contains the individual reseults for each file.
  \code{TRUE} if moved, \code{FALSE} if an error occured. Details of the eror
  re in the error files in the \code{input/bemovi} directory.
}
\description{
Check if data in input folder is OK and move to raw data folder
}
