% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tools_path.R
\name{check_tools_path}
\alias{check_tools_path}
\title{Checks tools_path() if all tools are there or downloads them}
\usage{
check_tools_path(path = tools_path(), download = FALSE)
}
\arguments{
\item{path}{path for the tools. defaults to \code{tools_path()}}

\item{download}{if \code{TRUE}, the tools are downloaded. If \code{FALSE},
the default, the ptools path is only checked.}
}
\value{
\code{TRUE} if all tools are installed. a named list of tools which re not installed.
}
\description{
Checks tools_path() if all tools are there or downloads them
}
\examples{
check_tools_path()

}
