% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify \code{morph_mvt} nad calculates densities}
\usage{
classify(
  bemovi_extract,
  morph_mvt,
  trajectory_data,
  classifiers_constant,
  classifiers_increasing,
  video_description_file,
  composition
)
}
\arguments{
\item{morph_mvt}{merged track data - one row per particle}

\item{classifiers_constant}{constant tempersture classifier}

\item{classifiers_increasing}{increasing temperature classifier}
}
\value{
\code{morph_mvt} with the classified species and probabilities
}
\description{
Classify \code{morph_mvt} nad calculates densities
}
