% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_LEEF_2.R
\name{classify_LEEF_2}
\alias{classify_LEEF_2}
\title{Classify \code{morph_mvt} nad calculates densities}
\usage{
classify_LEEF_2(
  bemovi_extract,
  morph_mvt,
  trajectory_data,
  classifiers,
  video_description_file,
  composition
)
}
\arguments{
\item{morph_mvt}{merged track data - one row per particle}

\item{classifiers}{list of classifiers TI (Temperature Increasing), NI (Nutrient Increasing), SI (Salt Increasing)}

\item{classifiers_constant}{constant temperature classifier}

\item{classifiers_increasing}{increasing temperature classifier}
}
\value{
\code{morph_mvt} with the classified species and probabilities
}
\description{
Classify \code{morph_mvt} nad calculates densities
}
