% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_bemovi.R
\name{extractor_bemovi}
\alias{extractor_bemovi}
\title{Extractor bemovi data}
\usage{
extractor_bemovi(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Analyse all \code{.avi} files in \code{bemovi} folder and save as \code{.rds} file.
}
\details{
This function is executing \code{extractor_bemovi_particle()},
\code{extractor_bemovi_trajectory()} and finally
\code{extractor_bemovi_merge()}
}
