% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_bemovi_classify.R
\name{extractor_bemovi_classify}
\alias{extractor_bemovi_classify}
\title{Extractor bemovi to id species and calculate density}
\usage{
extractor_bemovi_classify(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function id's the species and calculates the densities.
}
