% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_bemovi_filter.R
\name{extractor_bemovi_filter}
\alias{extractor_bemovi_filter}
\title{Extractor bemovi to filter traced particles}
\usage{
extractor_bemovi_filter(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function applies a filter.
}
