% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_bemovi.R
\name{extractor_bemovi_headless_linux}
\alias{extractor_bemovi_headless_linux}
\title{Extractor bemovi data FOR LINUX ON CLUSTER}
\usage{
extractor_bemovi_headless_linux(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Analyse all \code{.avi} files in \code{bemovi} folder and save as \code{.rds} file.
}
\details{
\bold{\code{extracxtor_bemovi_particle()} needs to be executed before!}

This function is executing
\code{extractor_bemovi_trajectory()} and finally
\code{extractor_bemovi_merge()}
}
