% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_bemovi_merge.R
\name{extractor_bemovi_merge}
\alias{extractor_bemovi_merge}
\title{Extractor bemovi data}
\usage{
extractor_bemovi_merge(input, output)
}
\arguments{
\item{input}{only for compatibility - not used}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Analyse all \code{.avi} files in \code{bemovi} folder and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database
(and therefore make accessible for further analysis and forecasting)
from \code{.avi} files.
}
