% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_bemovi_trajectory.R
\name{extractor_bemovi_trajectory}
\alias{extractor_bemovi_trajectory}
\title{Extractor bemovi data trajectories}
\usage{
extractor_bemovi_trajectory(input, output)
}
\arguments{
\item{input}{only for compatibility - not used}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function is calculating the trajectories
}
