% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_path.R
\name{tools_path}
\alias{tools_path}
\title{Read and write tools_path}
\usage{
tools_path(tools_path)
}
\arguments{
\item{tools_path}{\code{character} \code{vector} of length one containing the
directory where the tools are located}
}
\value{
tools directory. If values have set, the value before the change.
}
\description{
Read or write the tools path where the tools are located in. If no parameter
is given, the path will be returned as a list.
}
\examples{
tools_path()

tools_path("./temp")

}
