% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_bemovi_particle.R
\name{extractor_bemovi_particle}
\alias{extractor_bemovi_particle}
\title{Extractor bemovi data particles}
\usage{
extractor_bemovi_particle(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Extractor bemovi data particles
}
\details{
This function extracts the particles.
}
