% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processor_bemovi.R
\name{pre_processor_bemovi}
\alias{pre_processor_bemovi}
\title{Preprocessor bemovi data}
\usage{
pre_processor_bemovi(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
convert all \code{.cxd} files in \code{bemovi} folder to non-proprietory avi format and delete \code{.cxd} file.
}
